requirements.txt
absl-py==1.4.0
antlr4-python3-runtime==4.9.3
black==22.12.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
cloudpickle==2.2.1
contourpy==1.0.7
cycler==0.11.0
# -e git+https://github.com/facebookresearch/detectron2.git@58e472e076a5d861fdcf773d9254a3664e045bf8#egg=detectron2
Flask==2.2.2
fonttools==4.38.0
fvcore==0.1.5.post20221221
google-auth==2.16.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
hydra-core==1.3.1
idna==3.4
importlib-metadata==6.0.0
iopath==0.1.9
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
mypy-extensions==0.4.3
numpy==1.24.1
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.68
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==2.6.2
portalocker==2.7.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
tomli==2.0.1

#torch==1.13.1+cu116
#torchaudio==0.13.1+cu116
#torchvision==0.14.1+cu116

tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
yacs==0.1.8
zipp==3.11.0
